Moved mongodb.ts into /mongo folder
